requirements.txt
httpx==0.28.1
# --- FastAPI и сервер ---
fastapi==0.111.0
uvicorn[standard]==0.30.0

# --- OpenTelemetry ядро ---
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0

# --- Экспортёры OTLP (HTTP + gRPC) ---
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0

# --- Инструментализация ---
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-logging==0.48b0

# --- Логирование в OTEL/Vector ---
structlog==24.1.0